# openapi.yaml
openapi: 3.0.3
info:
  title: API MongoDB — Lista de Exercícios
  version: 1.0.0
  description: |-
    Lista de Exercícios de MongoDB implementada como API REST com **Node.js + Express**
    e o driver oficial `mongodb` (sem ODM).

    Cada grupo de rotas corresponde a uma parte da lista:

    | Recurso | Exercícios | Conceitos |
    |---|---|---|
    | Produtos | 2 a 11 | CRUD, `$gt`/`$lt`, `$and`/`$or`/`$nor`, projeção, `sort`/`limit`, `$set`, `$inc` |
    | Clientes | 12, 13, 14 | documentos aninhados, `$push`, `$pull`, `$addToSet` |
    | Filmes | 15 | modelagem com array + documento aninhado |
    | Alunos | 16 | arrays de valores, `$avg`, `$unwind` |
    | Vendas | 17, 18, 19 | `$match`, `$group`, `$project`, `$multiply`, `$sum`, `$avg` |
    | Pedidos | 20 | array de documentos, `$unwind`, agrupamento por status |
    | Exercícios | todos | enunciado + comando no mongosh + resultado do banco |

    A rota **`/api/exercicios`** devolve, para cada item da lista, o enunciado, o comando
    equivalente no `mongosh` e o resultado real vindo do banco.
  contact:
    name: Repositório do trabalho
    url: http://localhost:3000
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
components:
  schemas:
    ObjectId:
      type: string
      pattern: ^[0-9a-fA-F]{24}$
      description: Identificador único do documento (24 caracteres hexadecimais).
      example: 6a98c80f3e8030cb5731e645
...
    Venda:
      type: object
      required:
        - produto
        - categoria
        - quantidade
        - valorUnitario
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        produto:
          type: string
          example: Notebook
        categoria:
          type: string
          example: Informática
        quantidade:
          type: integer
          example: 2
        valorUnitario:
          type: number
          example: 4500
...
    ResultadoRemocao:
      type: object
      properties:
        removidos:
          type: integer
          example: 1
    ResultadoInsercaoMultipla:
      type: object
      properties:
        inseridos:
          type: integer
          example: 4
        ids:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ObjectId'
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: Recurso não encontrado.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: ObjectId do documento.
      schema:
        $ref: '#/components/schemas/ObjectId'
...
  responses:
    RequisicaoInvalida:
      description: Dados inválidos ou campos obrigatórios ausentes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    NaoEncontrado:
      description: Documento não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
...
paths:
  /api/vendas:
    get:
      tags:
        - Vendas
      summary: Lista vendas usando aggregate com $match e $sort (exercício 17)
      description: >
        O pipeline montado é devolvido junto da resposta, para conferência. O `$match` vem sempre antes do `$sort`,
        reduzindo o volume processado.
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
          example: Periféricos
          description: Estágio $match — exercício 17a.
        - name: quantidadeMin
          in: query
          schema:
            type: integer
          example: 3
          description: Estágio $match com $gt — exercício 17b.
        - name: ordenarPor
          in: query
          schema:
            type: string
            enum:
              - quantidade
              - valorUnitario
              - produto
        - name: ordem
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: Vendas e o pipeline utilizado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  pipeline:
                    type: array
                    items:
                      type: object
                    description: Estágios enviados ao MongoDB.
                  total:
                    type: integer
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Venda'
    post:
      tags:
        - Vendas
      summary: Registra uma ou várias vendas (exercício 17)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Venda'
                - type: array
                  items:
                    $ref: '#/components/schemas/Venda'
            example:
              produto: Notebook
              categoria: Informática
              quantidade: 2
              valorUnitario: 4500
      responses:
        '201':
          description: Venda(s) registrada(s).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoInsercaoMultipla'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
...
  /api/vendas/{id}:
    get:
      tags:
        - Vendas
      summary: Busca uma venda pelo _id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Venda encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venda'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
    delete:
      tags:
        - Vendas
      summary: Remove uma venda
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Documento removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoRemocao'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
...